% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{pm2alpha}
\alias{pm2alpha}
\title{Movement Probability to Alpha}
\usage{
pm2alpha(p)
}
\arguments{
\item{p}{Probability of moving.}
}
\value{
numeric vector.
}
\description{
Movement Probability to Alpha
}
\examples{

}
